Replace hmset/expire/hgetall with set/get
In the original implementation, there is a risk of an infinitely
long-lived cache key that arises if HMSET for the cache entry succeeds,
but EXPIRE fails or is not executed for any reason.

By serializing the cache `entry` to JSON and storing that single string
in Redis, we can take advantage of SET's ability to atomically accept an
EX-piration time along with the set.

The downside is we now have to do a JSON.stringify for every cache store
we do, and JSON.parse for every cache hit / retrieval, but that is
better than cache entries that never expire.

Tests pass against a local redis after modifying 1 test case which stores a
value then stores the same key again without cache. In the HMSET
implementation, this test passed because HMSET does not modify the
expiration date of existing keys. However, calls to SET always remove
old values and any expiration. Therefore, use a different key to store
the explicitly non-expiring cache entry.

Fixes #101.